Changes to be committed: 	modified:   .gitignore 	modified:   configs/pose_sample.yaml 	modified:   configs/pose_train.yaml 	modified:   configs/temp_train.yaml 	new file:   diffusers/ds_config.json 	modified:   diffusers/hf_data.py 	new file:   diffusers/inference.py 	new file:   diffusers/load_hf_data.py 	new file:   diffusers/run_config.yaml 	modified:   diffusers/train.sh 	modified:   diffusers/train_text_to_image.py 	new file:   ds_config.json 	modified:   followyourpose/models/unet.py 	new file:   followyourpose/models/unet_blocks_2d.py 	renamed:    followyourpose/models/unet_blocks.py -> followyourpose/models/unet_blocks_3d.py 	modified:   followyourpose/pipelines/pipeline_followyourpose.py 	new file:   inference.py 	modified:   multi_train.sh 	modified:   pose_example/metaworld_ButtonPress_20240826_205555_pose.mp4 	new file:   pose_example/metaworld_assembly-v2-goal-observable_20240926_004915_pose.mp4 	new file:   pose_example/pose.mp4 	new file:   run_config.yaml 	modified:   sample.sh 	modified:   single_gpu.sh 	new file:   test.py 	modified:   train_followyourpose.py 	modified:   txt2video.py

+++ followyourpose/pipelines/pipeline_followyourpose.py
@@ -378,7 +378,7 @@ class FollowYourPosePipeline(DiffusionPipeline):
         num_warmup_steps = len(timesteps) - num_inference_steps * self.scheduler.order
         
         skeleton, save_skeleton = self.get_skeleton(skeleton_path, video_length, frame_skeleton_stride)
-        skeleton = skeleton.to(latents.device).repeat(2,1,1,1,1)
+        skeleton = skeleton.to(latents.device).repeat(2,1,1,1,1).to(text_embeddings.dtype)
         
         with self.progress_bar(total=num_inference_steps) as progress_bar:
             for i, t in enumerate(timesteps):
